Moving cooperator lookup method to utils.py
